+++ telegrambots-meta/src/main/java/org/telegram/telegrambots/meta/api/methods/groupadministration/BanChatMember.java
@@ -1,0 +1,117 @@
+package org.telegram.telegrambots.meta.api.methods.groupadministration;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.core.type.TypeReference;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
+import org.telegram.telegrambots.meta.api.objects.ApiResponse;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiRequestException;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiValidationException;
+
+import java.io.IOException;
+import java.time.Duration;
+import java.time.Instant;
+import java.time.ZonedDateTime;
+
+/**
+ * @author Ruben Bermudez
+ * @version 5.3
+ * Use this method to kick a user from a group, a supergroup or a channel. In the case of supergroups,
+ * the user will not be able to return to the group on their own using invite links, etc., unless
+ * unbanned first. The bot must be an administrator in the group for this to work. Returns True on
+ * success.
+ * @apiNote  This will method only work if the ‘All Members Are Admins’ setting is off in the target
+ * group. Otherwise members may only be removed by the group's creator or by the member that added
+ * them.
+ */
+@EqualsAndHashCode(callSuper = false)
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@RequiredArgsConstructor
+@AllArgsConstructor
+@Builder
+public class BanChatMember extends BotApiMethod<Boolean> {
+    public static final String PATH = "banChatMember";
+
+    private static final String CHATID_FIELD = "chat_id";
+    private static final String USER_ID_FIELD = "user_id";
+    private static final String UNTILDATE_FIELD = "until_date";
+    private static final String REVOKEMESSAGES_FIELD = "revoke_messages";
+
+    @JsonProperty(CHATID_FIELD)
+    @NonNull
+    private String chatId; ///< Required. Unique identifier for the chat to send the message to (Or username for channels)
+    @JsonProperty(USER_ID_FIELD)
+    @NonNull
+    private Long userId; ///< Required. Unique identifier of the target user
+    @JsonProperty(UNTILDATE_FIELD)
+    private Integer untilDate; ///< Optional. Date when the user will be unbanned, unix time. If user is banned for more than 366 days or less than 30 seconds from the current time they are considered to be banned forever
+    /**
+     * Optional
+     *
+     * Pass True to delete all messages from the chat for the user that is being removed.
+     *
+     * If False, the user will be able to see messages in the group that were sent before the user was removed.
+     *
+     * Always True for supergroups and channels.
+     */
+    @JsonProperty(REVOKEMESSAGES_FIELD)
+    private Boolean revokeMessages;
+
+
+    @JsonIgnore
+    public void setUntilDateInstant(Instant instant) {
+        setUntilDate((int) instant.getEpochSecond());
+    }
+
+    @JsonIgnore
+    public void setUntilDateDateTime(ZonedDateTime date) {
+        setUntilDateInstant(date.toInstant());
+    }
+
+    @JsonIgnore
+    public void forTimePeriodDuration(Duration duration) {
+        setUntilDateInstant(Instant.now().plusMillis(duration.toMillis()));
+    }
+
+    @Override
+    public String getMethod() {
+        return PATH;
+    }
+
+    @Override
+    public Boolean deserializeResponse(String answer) throws TelegramApiRequestException {
+        try {
+            ApiResponse<Boolean> result = OBJECT_MAPPER.readValue(answer,
+                    new TypeReference<ApiResponse<Boolean>>(){});
+            if (result.getOk()) {
+                return result.getResult();
+            } else {
+                throw new TelegramApiRequestException("Error kicking chat member", result);
+            }
+        } catch (IOException e) {
+            throw new TelegramApiRequestException("Unable to deserialize response", e);
+        }
+    }
+
+    @Override
+    public void validate() throws TelegramApiValidationException {
+        if (chatId == null || chatId.isEmpty()) {
+            throw new TelegramApiValidationException("ChatId can't be empty", this);
+        }
+        if (userId == null || userId == 0) {
+            throw new TelegramApiValidationException("UserId can't be null or 0", this);
+        }
+    }
+}